Seperated Wonder Card component
Component to show wonder card under the board
